+++ src/python_apis/apis/__init__.py
@@ -38,12 +38,22 @@ Dependencies:
 
 """
 
-from python_apis.apis.ad_api import ADConnection
+from python_apis.apis.ad_api import (
+    ADConnection,
+    AD_COMPATIBILITY_ENV_VAR,
+    AD_COMPATIBILITY_MODES,
+    AD_DEFAULT_COMPATIBILITY_MODE,
+    resolve_ad_compatibility_mode,
+)
 from python_apis.apis.jira_api import JiraConnection
 from python_apis.apis.sql_api import SQLConnection
 
 __all__ = [
     "ADConnection",
+    "AD_COMPATIBILITY_ENV_VAR",
+    "AD_COMPATIBILITY_MODES",
+    "AD_DEFAULT_COMPATIBILITY_MODE",
+    "resolve_ad_compatibility_mode",
     "JiraConnection",
     "SQLConnection",
 ]